Moving NameTypeService to share/name-type/service folder.
Issue: IBP-4794
Review: none
