+++ src/main/jhipster/src/main/webapp/app/metadata-manager/metadata-manager.module.ts
@@ -5,7 +5,7 @@ import { METADATA_MANAGER_ROUTES } from './metadata-manager.route';
 import { MetadataManagerComponent } from './metadata-manager.component';
 import { NameTypeComponent } from './name-type/name-type.component';
 import { NameTypeResolvePagingParams } from './name-type/name-type-resolve-paging-params';
-import { NameTypeService } from './name-type/name-type.service';
+import { NameTypeService } from '../shared/name-type/service/name-type.service';
 import { NameTypeEditDialogComponent, NameTypeEditPopupComponent } from './name-type/name-type-edit-dialog.component';
 import { NameTypeContext } from './name-type/name-type.context';
